add auditorium-service add bookingticket-service add hall-service handle fallback feign client

+++ screening/src/main/java/com/whatamovie/screening/service/MovieService.java
@@ -11,5 +11,5 @@ import org.springframework.web.bind.annotation.RequestMapping;
 )
 public interface MovieService {
     @GetMapping("/api/v1/movie/{id}")
-    Movie findById(@PathVariable  int id);
+    Movie findById(@PathVariable  Long id);
 }
